+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The present file will list all changes made to the project; according to the
 #### Changes
 
 #### Deprecated
+- `Glpi\Inventory\Conf::importFile()`
 
 #### Removed